requirements.txt
# Dependencies
Flask==1.0.2
Flask-API==1.0
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.12

# Runtime
gunicorn==19.9.0
honcho==1.0.1

# Code quality
pylint==1.9.3
flake8==3.5.0

# Testing
factory_boy==2.11.1
mock==2.0.0
nose==1.3.7
pinocchio==0.4.2
rednose==1.3.0
httpie==0.9.9

# Code coverage
coverage==4.5.1
codecov==2.0.15
